requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
yfinance==0.2.40
joblib==1.4.2
scikit-learn==1.5.1
xgboost==2.1.1
pandas==2.3.3
numpy==1.26.4
lightgbm==4.4.0
fredapi==0.5.2
prophet==1.1.5
statsforecast==1.7.6
darts==0.30.0
ta==0.11.0
TA-Lib==0.4.28
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
requests==2.32.3
python-dotenv==1.0.0
shap==0.45.1
yellowbrick==1.5
jsonschema==4.20.0
pytest==7.4.0
pytest-cov==4.1.0
responses==0.23.1
pytest-httpx==0.23.0